:recycle: Refactor auth route into own file
