+++ README.md
@@ -2,8 +2,9 @@
 
 ### Commit Emoji Reference
 
-| Emoji       | Text          | What For                         |
-| ----------- | ------------- | -------------------------------- |
-| :blue_book: | `:blue_book:` | Commits with just specifications |
-| :books:     | `:books:`     | Commits with Documentation       |
-| :ledger:    | `:ledger:`    | Commits involving Tasks and Todo |
+| Emoji       | Text          | What For                            |
+| ----------- | ------------- | ----------------------------------- |
+| :blue_book: | `:blue_book:` | Commits with just specifications    |
+| :books:     | `:books:`     | Commits with Documentation          |
+| :ledger:    | `:ledger:`    | Commits involving Tasks and Todo    |
+| :recycle:   | `:recycle:`   | Green to Green Refactor Only commit |
