+++ src/client/mock_server/mock_server.rs
@@ -2,7 +2,8 @@ use url::Url;
 
 use client::TokenResponse;
 use client::mock_client::{MockReq, MockResp};
-use client::mock_server::{ServerResp, VALID_SCOPES};
+use client::mock_server::ServerResp;
+use client::mock_server::auth_route::auth;
 use client::params::UrlQueryParams;
 use errors::{Error, Result};
 
@@ -109,98 +110,16 @@ impl MockServer {
         }
     }
 
-    pub fn parse_state(url: &Url) -> Result<String> {
-        single_param("state", url)
-    }
-
-    pub fn parse_client_id(url: &Url) -> Result<String> {
-        let client_id = single_param("client_id", url)?;
-
-        if client_id != "someid@example.com" {
-            return Err(Error::unauthorized_client(
-                Some("Unauthorized: Client Not Authorized"),
-                None,
-            ));
-        }
-        Ok(client_id)
-    }
-
-    pub fn parse_redirect_uri(&self, url: &Url) -> Result<Option<Url>> {
-        match maybe_single_param("redirect_uri", url) {
-            Some(v) => Ok(validate_required_uri(v)?),
-            None => match self.redirect_uri_required {
-                true => {
-                    return Err(Error::invalid_request(
-                        Some("Bad Request: Missing `redirect_uri`"),
-                        None,
-                    ));
-                }
-                false => Ok(None),
-            },
-        }
-    }
-
-    pub fn parse_scope(&self, url: &Url) -> Result<()> {
-        let scope: Vec<String> = match UrlQueryParams::from(url.query_pairs()).get("scope") {
-            Some(v) => v.into_iter().collect(),
-            None => vec![],
-        };
-
-        for value in scope {
-            if !VALID_SCOPES.into_iter().any(|&v| v == value) {
-                return Err(Error::invalid_request(
-                    None,
-                    Some(format!(
-                        "https://docs.example.com/scopes?invalid_scope={}",
-                        value
-                    )),
-                ));
-            }
-        }
-
-        Ok(())
-    }
-
-    pub fn auth(&self, req: MockReq) -> ServerResp {
-        if let Some(ref err) = self.error {
-            return ServerResp::redirect_err(err);
-        }
-        let state = match MockServer::parse_state(&req.url) {
-            Ok(k) => k,
-            Err(e) => return ServerResp::redirect_err(&e),
-        };
-
-        match MockServer::parse_client_id(&req.url) {
-            Ok(_) => (),
-            Err(e) => return ServerResp::redirect_err(&e),
-        };
-
-        match self.parse_redirect_uri(&req.url) {
-            Ok(_) => (),
-            Err(e) => return ServerResp::redirect_err(&e),
-        };
-
-        match self.parse_scope(&req.url) {
-            Ok(_) => (),
-            Err(e) => return ServerResp::redirect_err(&e),
-        };
-
-        Ok(MockReq {
-            url: Url::parse_with_params(
-                "https://localhost/example/auth",
-                vec![("state", state), ("code", "MOCK_CODE".into())],
-            ).unwrap(),
-            body: String::from(""),
-        }).into()
-    }
-
     pub fn token(&self, _: MockReq) -> ServerResp {
-        MockResp::parse_access_token_response(&TokenResponse::new("2YotnFZFEjr1zCsicMWpAA", "bearer")).into()
+        MockResp::parse_access_token_response(&TokenResponse::new(
+            "2YotnFZFEjr1zCsicMWpAA",
+            "bearer",
+        )).into()
     }
 
     pub fn send_request(&self, req: MockReq) -> ServerResp {
         match req.url.path() {
-            "/auth" => self.auth(req),
+            "/auth" => auth(self, req),
             "/token" => self.token(req),
             _ => ServerResp::response_err(&Error::msg("404: Route not found")),
         }
